fix(layout): prevent redirect on missing session and return null instead refactor(header): update import path for user-profile component and remove unnecessary prop
